Autopilot_light/src/requirements_n.txt
# Should contain the required packages for the code 2020-05-04
argh
certifi
cycler
fire
kiwisolver
libais
matplotlib
numpy
pybind11
pybind11-cmake
pynmea2
pyparsing
pyserial
python-dateutil
scipy
six
termcolor
